Unified setting of accepting invalid SSL certificates in Web Auth project
