+++ powerauth-webauth-client/src/main/resources/application.properties
@@ -7,7 +7,8 @@ powerauth.webauth.service.oauth.clientSecret=changeme
 
 # Next Step Server Service URL
 powerauth.nextstep.service.url=http://localhost:8080/powerauth-nextstep
-powerauth.nextstep.service.ssl.acceptInvalidSslCertificate=false
+# Whether invalid SSL certificates should be accepted
+powerauth.service.ssl.acceptInvalidSslCertificate=false
 
 # Database Keep-Alive
 spring.datasource.test-while-idle=true
