+++ powerauth-webauth/src/main/java/io/getlime/security/powerauth/app/webauth/configuration/PowerAuthWebServiceConfiguration.java
@@ -1,8 +1,10 @@
 package io.getlime.security.powerauth.app.webauth.configuration;
 
 import io.getlime.push.client.PushServerClient;
+import io.getlime.security.powerauth.lib.webauth.authentication.service.SSLConfigurationService;
 import io.getlime.security.powerauth.soap.spring.client.PowerAuthServiceClient;
 import org.apache.ws.security.WSConstants;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -10,13 +12,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.oxm.jaxb.Jaxb2Marshaller;
 import org.springframework.ws.client.support.interceptor.ClientInterceptor;
 import org.springframework.ws.soap.security.wss4j.Wss4jSecurityInterceptor;
-
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * Configuration for the PowerAuth 2.0 Server connector.
@@ -26,6 +21,8 @@ import java.util.logging.Logger;
 @Configuration
 @ComponentScan(basePackages = {"io.getlime.security.powerauth"})
 public class PowerAuthWebServiceConfiguration {
+
+    private SSLConfigurationService sslConfigurationService;
 
     @Value("${powerauth.service.url}")
     private String powerAuthServiceUrl;
@@ -41,6 +38,11 @@ public class PowerAuthWebServiceConfiguration {
 
     @Value("${powerauth.service.ssl.acceptInvalidSslCertificate}")
     private boolean acceptInvalidSslCertificate;
+
+    @Autowired
+    public PowerAuthWebServiceConfiguration(SSLConfigurationService sslConfigurationService) {
+        this.sslConfigurationService = sslConfigurationService;
+    }
 
     // Must use DEPRECATED class here, wss4j2 is not yet production ready
     @Bean
@@ -72,7 +74,7 @@ public class PowerAuthWebServiceConfiguration {
         }
         // whether invalid SSL certificates should be accepted
         if (acceptInvalidSslCertificate) {
-            trustAllCertificates();
+            sslConfigurationService.trustAllCertificates();
         }
         return client;
     }
@@ -83,42 +85,9 @@ public class PowerAuthWebServiceConfiguration {
         client.setServiceBaseUrl(powerAuthPushServiceUrl);
         // whether invalid SSL certificates should be accepted
         if (acceptInvalidSslCertificate) {
-            trustAllCertificates();
+            sslConfigurationService.trustAllCertificates();
         }
         return client;
-    }
-
-    /**
-     * Activate trust in all SSL certificates including invalid ones for non-production use.
-     */
-    private void trustAllCertificates() {
-        HttpsURLConnection.setDefaultHostnameVerifier((hostname, session) -> true);
-
-        TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
-
-            public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-                return null;
-            }
-
-            public void checkClientTrusted(java.security.cert.X509Certificate[] certs, String authType) {
-            }
-
-            public void checkServerTrusted(java.security.cert.X509Certificate[] certs, String authType) {
-            }
-
-        }};
-
-        try {
-            SSLContext sc = SSLContext.getInstance("SSL");
-            sc.init(null, trustAllCerts, new java.security.SecureRandom());
-            HttpsURLConnection.setDefaultSSLSocketFactory(sc.getSocketFactory());
-        } catch (Exception e) {
-            Logger.getLogger(this.getClass().getName()).log(
-                    Level.SEVERE,
-                    "Error occurred while setting SSL socket factory",
-                    e
-            );
-        }
     }
 
 }
