+++ powerauth-webauth/src/main/resources/application.properties
@@ -1,15 +1,14 @@
 # Credential Server Service URL
 powerauth.credentials.service.url=http://localhost:8080/powerauth-credential-server-sample
-powerauth.credentials.service.ssl.acceptInvalidSslCertificate=false
 
 # Next Step Server Service URL
 powerauth.nextstep.service.url=http://localhost:8080/powerauth-nextstep
-powerauth.nextstep.service.ssl.acceptInvalidSslCertificate=false
 
 # PowerAuth 2.0 Server URL
 powerauth.service.url=http://localhost:8080/powerauth-java-server/soap
 powerauth.service.security.clientToken=
 powerauth.service.security.clientSecret=
+# Whether invalid SSL certificates should be accepted
 powerauth.service.ssl.acceptInvalidSslCertificate=false
 
 # PowerAuth 2.0 Push Server URL
